Try harder to remove unused shared memory files
This commit introduces a shared flock on the shm file which
is held by the client and the server. In addition to checking
if the shm file already exists, we introduce a recovery code
path which uses an exclusive flock to find out if a process is
still holding the file descriptor. If not, we may try to delete
it and re-try creating a fresh shm file again.

Change-Id: I79150953acb3f0d8801fa0f254b3d9ac20fc2f4d