Fix ARQStream claiming no packets available
Received packets are placed in a shared memory region by the HostARQ
daemon. The user API accesses this shared memory. A user-API-local
buffer (`recv_buf`) is used to prefetch up to one cacheline worth of
pointers to those packets.

The received_packet_available() function, however, only checked the
state of the input fifo and NOT the local buffer.

As long as the client consumes the frames faster than the ARQ can decode
them, there is only one frame per FIFO-entry. If the client stalls,
however, there will be up to eight, at which point the last batch will
never be served because it is in the local buffer, but
received_packet_available() claims there are no new packets to serve.

Drive-By: Adjust order of function definitions to be consistent with
header file.

Change-Id: I59af2b2522959230e3eff2f8acbfb8c78e8cc46f