Fix ARQStream claiming all packets are sent
or "Add tx_send_buf_empty and use in ARQStream<P>::all_packets_sent()"
(follow-up to c/12042).

Packets to be sent are placed in a shared memory region by the HostARQ
daemon. The user API accesses this shared memory. A user-API-local
buffer (`send_buf`) is used to prefill up to one cacheline worth of
pointers to those packets (or buffers, in sctrltp terms).

The all_packets_sent() function, however, only checked the state of the
output queue/fifo and NOT the local send buffer.

This change won't have effects in most existing use cases as we usually
use MODE_FLUSH that triggers flushing of the local send buffer.

Change-Id: I28aa591be022bf06b57ee853fafc2d8683a6e404